# api/openapi.yaml
openapi: 3.0.3
info:
  title: UptimeX Code API
  description: API for UptimeX Code - AI coding agent for DevOps/SRE
  version: 0.0.1
  contact:
    name: UptimeX
    url: https://uptimex.cloud

servers:
  - url: https://api2.uptimex.cloud/api/v1
    description: Production
  - url: http://localhost:8080/api/v1
    description: Local development

paths:
  /health:
    get:
      summary: Health check
      responses:
        '200':
          description: OK

  /auth/login:
    post:
      summary: Login
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                provider:
                  type: string
                  enum: [github, google, email]
                email:
                  type: string
      responses:
        '200':
          description: Login successful

  /user/profile:
    get:
      summary: Get user profile
      tags: [User]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

  /chat/completions:
    post:
      summary: Create chat completion
      tags: [Chat]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
      responses:
        '200':
          description: Chat completion response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer

  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        plan:
          type: string
          enum: [free, pro, enterprise]

    ChatRequest:
      type: object
      required:
        - messages
      properties:
        model:
          type: string
        messages:
          type: array
          items:
            $ref: '#/components/schemas/Message'
        stream:
          type: boolean

    Message:
      type: object
      properties:
        role:
          type: string
          enum: [system, user, assistant]
        content:
          type: string

    ChatResponse:
      type: object
      properties:
        id:
          type: string
        choices:
          type: array
          items:
            type: object
            properties:
              message:
                $ref: '#/components/schemas/Message'